refactor: improve package separation
