+++ lib/src/handshake/exception.dart
