+++ lib/src/parallelism/exception.dart
